requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
contourpy==1.3.0
cycler==0.12.1
distro==1.9.0
fonttools==4.53.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.7
jiter==0.5.0
joblib==1.4.2
kiwisolver==1.4.6
matplotlib==3.9.2
numpy==2.1.0
openai==1.42.0
packaging==24.1
pillow==10.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
scikit-learn==1.5.1
scipy==1.14.1
six==1.16.0
sniffio==1.3.1
threadpoolctl==3.5.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.2
